# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: merchants
    description: Merchant operations
  - name: votes
    description: Vote operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /merchants:
    get:
      operationId: getMerchants
      tags: [merchants]
      summary: Get all merchants
      description: Returns a list of all merchants with their trust scores
      responses:
        "200":
          description: List of merchants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Merchant"

    post:
      operationId: createMerchant
      tags: [merchants]
      summary: Create a merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMerchantBody"
      responses:
        "201":
          description: Created merchant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"

  /merchants/{upiId}:
    get:
      operationId: getMerchant
      tags: [merchants]
      summary: Get merchant by UPI ID
      description: Returns merchant trust info and risk level
      parameters:
        - name: upiId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "404":
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /merchants/{upiId}/vote:
    post:
      operationId: submitVote
      tags: [votes]
      summary: Submit a post-payment vote
      description: Submit vote after transaction. Anti-fake logic applied.
      parameters:
        - name: upiId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVoteBody"
      responses:
        "200":
          description: Vote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "400":
          description: Already voted or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /merchants/{upiId}/stats:
    get:
      operationId: getMerchantStats
      tags: [merchants]
      summary: Get detailed merchant stats
      description: Returns full analytics including weekly breakdown and vote history
      parameters:
        - name: upiId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantStats"
        "404":
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /merchants/{upiId}/report-fraud:
    post:
      operationId: reportFraud
      tags: [merchants]
      summary: Report merchant as fraudulent
      description: Fraud reports reduce trust score faster
      parameters:
        - name: upiId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFraudBody"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudReportResult"
        "404":
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [merchants]
      summary: Get dashboard summary stats
      description: Returns platform-wide trust stats for the dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Merchant:
      type: object
      properties:
        id:
          type: integer
        upiId:
          type: string
        name:
          type: string
        category:
          type: string
        totalTransactions:
          type: integer
        happyTransactions:
          type: integer
        fraudReports:
          type: integer
        trustScore:
          type: number
        satisfactionPercent:
          type: number
        riskLevel:
          type: string
          enum: [Safe, Medium, Risky]
        isVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - upiId
        - name
        - category
        - totalTransactions
        - happyTransactions
        - fraudReports
        - trustScore
        - satisfactionPercent
        - riskLevel
        - isVerified
        - createdAt

    CreateMerchantBody:
      type: object
      properties:
        upiId:
          type: string
        name:
          type: string
        category:
          type: string
        isVerified:
          type: boolean
      required:
        - upiId
        - name
        - category

    SubmitVoteBody:
      type: object
      properties:
        voterUpiId:
          type: string
        isHappy:
          type: boolean
        isTransactionSafe:
          type: boolean
        didMerchantBehave:
          type: boolean
        isSatisfied:
          type: boolean
        amountPaid:
          type: ["number", "null"]
      required:
        - voterUpiId
        - isHappy
        - isTransactionSafe
        - didMerchantBehave
        - isSatisfied

    VoteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        updatedMerchant:
          $ref: "#/components/schemas/Merchant"
      required:
        - success
        - message
        - updatedMerchant

    ReportFraudBody:
      type: object
      properties:
        reporterUpiId:
          type: string
        reason:
          type: string
      required:
        - reporterUpiId
        - reason

    FraudReportResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    MerchantStats:
      type: object
      properties:
        merchant:
          $ref: "#/components/schemas/Merchant"
        weeklyVotes:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyVoteStat"
        recentVotes:
          type: array
          items:
            $ref: "#/components/schemas/VoteRecord"
        fraudReportCount:
          type: integer
      required:
        - merchant
        - weeklyVotes
        - recentVotes
        - fraudReportCount

    WeeklyVoteStat:
      type: object
      properties:
        week:
          type: string
        happy:
          type: integer
        total:
          type: integer
      required:
        - week
        - happy
        - total

    VoteRecord:
      type: object
      properties:
        id:
          type: integer
        isHappy:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - isHappy
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalMerchants:
          type: integer
        totalTransactions:
          type: integer
        safeMerchants:
          type: integer
        mediumMerchants:
          type: integer
        riskyMerchants:
          type: integer
        averageTrustScore:
          type: number
        totalFraudReports:
          type: integer
        topTrustedMerchants:
          type: array
          items:
            $ref: "#/components/schemas/Merchant"
      required:
        - totalMerchants
        - totalTransactions
        - safeMerchants
        - mediumMerchants
        - riskyMerchants
        - averageTrustScore
        - totalFraudReports
        - topTrustedMerchants

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error