# lib/api-zod/src/generated/types/createMerchantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMerchantBody {
  upiId: string;
  name: string;
  category: string;
  isVerified?: boolean;
}