# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Merchant } from "./merchant";

export interface DashboardSummary {
  totalMerchants: number;
  totalTransactions: number;
  safeMerchants: number;
  mediumMerchants: number;
  riskyMerchants: number;
  averageTrustScore: number;
  totalFraudReports: number;
  topTrustedMerchants: Merchant[];
}