# lib/api-zod/src/generated/types/merchant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MerchantRiskLevel } from "./merchantRiskLevel";

export interface Merchant {
  id: number;
  upiId: string;
  name: string;
  category: string;
  totalTransactions: number;
  happyTransactions: number;
  fraudReports: number;
  trustScore: number;
  satisfactionPercent: number;
  riskLevel: MerchantRiskLevel;
  isVerified: boolean;
  createdAt: string;
}