# lib/api-zod/src/generated/types/merchantStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Merchant } from "./merchant";
import type { VoteRecord } from "./voteRecord";
import type { WeeklyVoteStat } from "./weeklyVoteStat";

export interface MerchantStats {
  merchant: Merchant;
  weeklyVotes: WeeklyVoteStat[];
  recentVotes: VoteRecord[];
  fraudReportCount: number;
}