# lib/api-zod/src/generated/types/reportFraudBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportFraudBody {
  reporterUpiId: string;
  reason: string;
}