# lib/api-zod/src/generated/types/submitVoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitVoteBody {
  voterUpiId: string;
  isHappy: boolean;
  isTransactionSafe: boolean;
  didMerchantBehave: boolean;
  isSatisfied: boolean;
  /** @nullable */
  amountPaid?: number | null;
}