# lib/api-zod/src/generated/types/voteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Merchant } from "./merchant";

export interface VoteResult {
  success: boolean;
  message: string;
  updatedMerchant: Merchant;
}